Add diffing adapter
