+++ epoxy-processor/src/main/java/com/airbnb/epoxy/ProcessorUtils.java
@@ -18,10 +18,15 @@ import javax.lang.model.type.TypeKind;
 import javax.lang.model.type.TypeMirror;
 import javax.lang.model.util.Types;
 
+import static javax.lang.model.element.ElementKind.CLASS;
+import static javax.lang.model.element.Modifier.PRIVATE;
+import static javax.lang.model.element.Modifier.STATIC;
+
 class ProcessorUtils {
 
   static final String EPOXY_MODEL_TYPE = "com.airbnb.epoxy.EpoxyModel<?>";
   static final String EPOXY_MODEL_HOLDER_TYPE = "com.airbnb.epoxy.EpoxyModelWithHolder<?>";
+  static final String EPOXY_DIFF_ADAPTER_TYPE = "com.airbnb.epoxy.DiffAdapter";
 
   static void throwError(String msg, Object... args)
       throws EpoxyProcessorException {
@@ -38,6 +43,10 @@ class ProcessorUtils {
 
   static boolean isEpoxyModel(TypeMirror type) {
     return isSubtypeOfType(type, EPOXY_MODEL_TYPE);
+  }
+
+  static boolean isDiffAdapter(TypeElement element) {
+    return isSubtypeOfType(element.asType(), EPOXY_DIFF_ADAPTER_TYPE);
   }
 
   static boolean isEpoxyModel(TypeElement type) {
@@ -183,5 +192,45 @@ class ProcessorUtils {
     // TODO: (eli_hart 2/2/17) If the class added additional types this won't be correct. That
     // should be rare, but it would be nice to handle.
     return superTypeArguments.get(0);
+  }
+
+  static void validateFieldAccessibleViaGeneratedCode(Element fieldElement,
+      Class<?> annotationClass, ErrorLogger errorLogger) {
+    TypeElement enclosingElement = (TypeElement) fieldElement.getEnclosingElement();
+
+    // Verify method modifiers.
+    Set<Modifier> modifiers = fieldElement.getModifiers();
+    if (modifiers.contains(PRIVATE) || modifiers.contains(STATIC)) {
+      errorLogger.logError(
+          "%s annotations must not be on private or static fields. (class: %s, field: %s)",
+          annotationClass.getSimpleName(),
+          enclosingElement.getSimpleName(), fieldElement.getSimpleName());
+    }
+
+    // Nested classes must be static
+    if (enclosingElement.getNestingKind().isNested()) {
+      if (!enclosingElement.getModifiers().contains(STATIC)) {
+        errorLogger.logError(
+            "Nested classes with %s annotations must be static. (class: %s, field: %s)",
+            annotationClass.getSimpleName(),
+            enclosingElement.getSimpleName(), fieldElement.getSimpleName());
+      }
+    }
+
+    // Verify containing type.
+    if (enclosingElement.getKind() != CLASS) {
+      errorLogger
+          .logError("%s annotations may only be contained in classes. (class: %s, field: %s)",
+              annotationClass.getSimpleName(),
+              enclosingElement.getSimpleName(), fieldElement.getSimpleName());
+    }
+
+    // Verify containing class visibility is not private.
+    if (enclosingElement.getModifiers().contains(PRIVATE)) {
+      errorLogger.logError(
+          "%s annotations may not be contained in private classes. (class: %s, field: %s)",
+          annotationClass.getSimpleName(),
+          enclosingElement.getSimpleName(), fieldElement.getSimpleName());
+    }
   }
 }
