+++ epoxy-processortest/src/test/resources/AdapterWithAutoModel_EpoxyHelper.java
@@ -1,0 +1,15 @@
+package com.airbnb.epoxy.adapter;
+
+import com.airbnb.epoxy.AdapterHelper;
+import com.airbnb.epoxy.BasicModelWithAttribute_;
+import java.lang.Override;
+
+/**
+ * Generated file. Do not modify!
+ */
+public class AdapterWithAutoModel_EpoxyHelper extends AdapterHelper<AdapterWithAutoModel> {
+  @Override
+  public void buildAutoModels(AdapterWithAutoModel adapter) {
+    adapter.modelWithAttribute_ = new BasicModelWithAttribute_().id(-1);
+  }
+}
