+++ epoxy-sample/src/main/java/com/airbnb/epoxy/models/ColorModel.java
@@ -3,6 +3,7 @@ package com.airbnb.epoxy.models;
 import android.support.annotation.ColorInt;
 import android.view.View;
 
+import com.airbnb.epoxy.ColorData;
 import com.airbnb.epoxy.EpoxyAttribute;
 import com.airbnb.epoxy.EpoxyModel;
 import com.airbnb.epoxy.EpoxyModelClass;
@@ -18,8 +19,9 @@ import com.airbnb.epoxy.R;
 public abstract class ColorModel extends EpoxyModel<View> {
   @EpoxyAttribute @ColorInt int color;
 
-  public ColorModel(@ColorInt int color) {
-    this.color = color;
+  public ColorModel(ColorData color) {
+    this.color = color.getColorInt();
+    id(color.getId());
   }
 
   @Override
